code is modularized
