+++ public/index.html
@@ -33,13 +33,13 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Git Set Go</title>
 </head>
 
-<body style="background-image: url(bg-main.jpg);  background-size: cover;">
+<body style="background-image: url(./images/bg-main.jpg);  background-size: cover;">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <a href="https://github.com/mrinal41298/gitsetgo"><img src="forkme_right_red.svg" style="position:absolute;top:0;right:0;" alt="Fork me on GitHub"></a>
+    <a href="https://github.com/mrinal41298/gitsetgo"><img src="./images/forkme_right_red.svg" style="position:absolute;top:0;right:0;" alt="Fork me on GitHub"></a>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
